+++ src/components/switch/switch.test.ts
@@ -13,7 +13,7 @@ describe('<wa-switch>', async () => {
   it('default properties', async () => {
     const el = await fixture<WaSwitch>(html` <wa-switch></wa-switch> `);
 
-    expect(el.name).to.equal('');
+    expect(el.name).to.equal(null);
     expect(el.value).to.be.null;
     expect(el.title).to.equal('');
     expect(el.disabled).to.be.false;