+++ src/components/textarea/textarea.test.ts
@@ -15,7 +15,7 @@ describe('<wa-textarea>', async () => {
     const el = await fixture<WaTextarea>(html` <wa-textarea></wa-textarea> `);
 
     expect(el.size).to.equal('medium');
-    expect(el.name).to.equal('');
+    expect(el.name).to.equal(null);
     expect(el.value).to.equal('');
     expect(el.defaultValue).to.equal('');
     expect(el.title).to.equal('');